Remove Housekeeping implementation from CT
