+++ circus-train-housekeeping/src/test/resources/com/hotels/housekeeping/repository/legacyReplicaPath-after-save-test.xml
