+++ circus-train-integration-tests/src/test/java/com/hotels/bdp/circustrain/integration/CircusTrainHdfsHdfsIntegrationTest.java
@@ -78,9 +78,9 @@ import com.google.common.collect.FluentIterable;
 
 import com.hotels.bdp.circustrain.common.test.base.CircusTrainRunner;
 import com.hotels.bdp.circustrain.common.test.junit.rules.ServerSocketRule;
-import com.hotels.bdp.circustrain.housekeeping.model.LegacyReplicaPath;
 import com.hotels.bdp.circustrain.integration.utils.TestUtils;
 import com.hotels.beeju.ThriftHiveMetaStoreJUnitRule;
+import com.hotels.housekeeping.model.LegacyReplicaPath;
 
 public class CircusTrainHdfsHdfsIntegrationTest {
   private static final Logger LOG = LoggerFactory.getLogger(CircusTrainHdfsHdfsIntegrationTest.class);
