+++ circus-train-integration-tests/src/test/java/com/hotels/bdp/circustrain/integration/utils/TestUtils.java
@@ -53,7 +53,7 @@ import com.google.common.base.Joiner;
 import com.google.common.collect.FluentIterable;
 import com.google.common.collect.ImmutableList;
 
-import com.hotels.bdp.circustrain.housekeeping.model.LegacyReplicaPath;
+import com.hotels.housekeeping.model.LegacyReplicaPath;
 
 public final class TestUtils {
