+++ circus-train-tool-parent/circus-train-vacuum-tool/src/test/java/com/hotels/bdp/circustrain/tool/vacuum/VacuumToolApplicationTest.java
@@ -52,9 +52,9 @@ import com.hotels.bdp.circustrain.api.metastore.CloseableMetaStoreClient;
 import com.hotels.bdp.circustrain.core.conf.ReplicaTable;
 import com.hotels.bdp.circustrain.core.conf.TableReplication;
 import com.hotels.bdp.circustrain.core.conf.TableReplications;
-import com.hotels.bdp.circustrain.housekeeping.model.LegacyReplicaPath;
-import com.hotels.bdp.circustrain.housekeeping.repository.LegacyReplicaPathRepository;
-import com.hotels.bdp.circustrain.housekeeping.service.HousekeepingService;
+import com.hotels.housekeeping.model.LegacyReplicaPath;
+import com.hotels.housekeeping.repository.LegacyReplicaPathRepository;
+import com.hotels.housekeeping.service.HousekeepingService;
 
 @RunWith(MockitoJUnitRunner.class)
 public class VacuumToolApplicationTest {
